v2.0.0-beta1
v2.0.0-beta1

Here is a list of the NVIDIA Shield models currently available to the
public: (let me know if I'm missing one)
 - Shield TV Pro (2019)
 - Shield TV (2019)
 - Shield TV Pro (2017)
 - Shield TV (2017)
 - Shield TV Pro (2015)
 - Shield TV (2015)

- Remote Style Override: Regardless of the kind of device you own, you
can switch your Firemote card to look like a completely different remote
control.  If you were very clever, you probably knew of a way to do this
already, but now it's built in!
- Visible Device Name Color Choice: As requested in #44, you can now
choose a different color for the visible device name text.  As of right
now, it should work properly on any Amazon style remote, but it is not
yet ready for the Shield version.
- Sometimes apps made for the Amazon Fire world weren't the same as the
apps made for other Android TV style devices which needed to be sorted
out.  If you've used Firemote for any non Amazon Fire devices, you may
have come across this problem before, as @moh-Brent discovered in #30.
These issues have been all sorted out under the hood which will
absolutely result in a better experience

- Navigation via the dpad is sllloooowwww.  While I recognize that this
is the common experience for anyone who has been using Firemote with a
Shield device in 'Strong' compatibility mode, I really want to fix it.
So far, lots of research, no positive results.
- Support for the visible device name feature on an Shield remote has
not been completed, so attempting to use it now will completely kill the
layout of an NVIDIA style remote