Xiaomi Mi Box Beta 3
Changes:
- Default event changed for Mi Box S from event3 to event2.  This could
have been the cause of the report from @moh-Brent in #55 that the
"apps-button" was not doing anything
- On the topic of the apps button not working, I replaced the Xiaomi
command tied to that button with an adb start command which does not
require a proper "event" number to be selected, and is faster than a
"Strong" config.  Hopefully I've double solved this issue.
- Adjusted the launch command for the emby app as reported by
@ChristophCaina in #18 to be problematic.  This change should make it
more stable and even faster for everyone who uses it.
- Adjusted the launch command for Magenta TV as reported by
@ChristophCaina in #58.  This needs testing/confirmation.
- Added a confirmation dialog before executing a function: restart
button press command as noted by @moh-Brent in #55 (good call Kiwi)

Not finished yet:
 - nothing!  (let me know if I missed something)